requirements.txt
aiofiles==24.1.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dataclasses-json==0.6.7
emoji==2.14.0
eval_type_backport==0.2.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
filetype==1.2.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.21
langchain-google-genai==2.0.5
langdetect==1.0.9
langsmith==0.1.145
lxml==5.3.0
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.14.2
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.3
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.40.1
tabulate==0.9.0
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.0
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.11.8
unstructured-client==0.28.0
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.45.1
wrapt==1.17.0
